requirements.txt
Flask==3.0.3
torch==2.5.0
pandas==2.2.3
open-clip-torch==2.29.0
